pyproject.toml
[tool.poetry]
name = "fadmin"
version = "0.0.0"
description = "Factorio server management via RCON"
authors = ["Truman Kilen <tkilen@mm.st>"]

[tool.poetry.scripts]
fadmin = "fadmin:main"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.0"
python-dotenv = "^0.10.1"
factorio-rcon-py = "^0.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
